Bump version for release
Interestingly, this is a perfect example of the kind of change Rich Hickey
talked about in his talk "Maybe Not", where a dynamically typed language can
extend the types handled by a function without breaking existing usage. In a
strongly typed language it would take a lot of incidental complexity to
avoid (and never completely) having to visit every call site and change the
types of arguments.

I wish I had time to try to figure out the best, most flexible way to handle
this kind of thing in Rust or some other strongly typed language.